+++ lib/npm/browser.ts
@@ -45,6 +45,7 @@ export const analyzeMetafileSync: typeof types.analyzeMetafileSync = () => {
 
 export const stop = () => {
   if (stopService) stopService()
+  return Promise.resolve()
 }
 
 interface Service {